refactor(frontend/menu): refactor menu structure components - remove current-stage-phase for the benefit of client-content/ current-view etc - add options to menu (empty for now)
