+++ frontend/src/components/home/client-content.module.scss
@@ -1,7 +1,7 @@
 $maxHeightNavigation: 5%;
 $minHeightContent: 95%;
 
-.currentStagePhaseWrapper {
+.clientContentWrapper {
   height: 100%;
   width: 100%;
