+++ frontend/src/components/home/home-content.tsx
@@ -1,13 +1,14 @@
 "use client";
 import { FC } from "react";
 import styles from "./home-content.module.scss";
-import { CurrentPhaseStage, Sidebar } from "@/components";
+import { Sidebar } from "@/components";
+import { ClientContent } from "./client-content";
 
 export const HomeContent: FC = () => {
   return (
     <div className={styles.homeContentWrapper}>
       <div className={styles.clientContent}>
-        <CurrentPhaseStage />
+        <ClientContent />
       </div>
       <div className={styles.sidebar}>
         <Sidebar />
